Rendering a TEXTURE! + Refactoring
Renders the player as a Sprite instead of just a filled rectangle. Now for the rest of the objects!

craeted SDLData class to hold all of the semi-global SDL stuff that needs to stay alive throughout the game, to be destroyed at close-time.

Created a Sprite class that will be responsible for drawing and positioning objects in the world, including stretching and rotating.

Refactoring of Main and Types to support these changes.